Revert "Merchant edit dish"
