+++ src/app/customer/customer.component.html
@@ -633,7 +633,7 @@
               </div>
             </div>
           </div>
-<!--          <div class="col-md-12 text-center load-more">-->
+          <div class="col-md-12 text-center load-more">
             <button class="btn btn-primary" type="button" disabled>
               <span class="spinner-grow spinner-grow-sm" role="status" aria-hidden="true"></span>
               Loading...
@@ -642,6 +642,7 @@
         </div>
       </div>
     </div>
+  </div>
 </section>
 <app-most-purchased-dishes></app-most-purchased-dishes>
 <section class="section pt-5 pb-5 text-center bg-white">
