+++ src/app/customer/detail-food/detail-food.component.ts
@@ -1,8 +1,11 @@
 import {Component, OnInit} from '@angular/core';
 import {Dish} from '../../model/dish';
 import {DishService} from '../../service/dish/dish.service';
-import {ActivatedRoute, ParamMap} from '@angular/router';
-import {Merchant} from '../../model/merchant';
+import {ActivatedRoute, ParamMap, Router} from '@angular/router';
+import {CartService} from '../../service/cart/cart.service';
+import {NotificationService} from '../../service/notification/notification.service';
+import {Cart} from '../../model/cart';
+import {AuthService} from '../../service/auth/auth.service';
 
 @Component({
   selector: 'app-detail-food',
@@ -12,12 +15,18 @@ import {Merchant} from '../../model/merchant';
 export class DetailFoodComponent implements OnInit {
 
   id: number;
-  merchant: Merchant;
-
-  refreshNum = 0;
+  dish: Dish;
+  dishTotal: number;
+  dishImage: string;
+  quantity: number;
+  cart: Cart;
 
   constructor(private dishService: DishService,
               private activatedRoute: ActivatedRoute,
+              private cartService: CartService,
+              private notificationService: NotificationService,
+              private router: Router,
+              private authService: AuthService
   ) {
     this.activatedRoute.paramMap.subscribe((paramMap: ParamMap) => {
       this.id = +paramMap.get('id');
@@ -26,19 +35,65 @@ export class DetailFoodComponent implements OnInit {
 
   ngOnInit() {
     document.getElementById('dish-image').scrollIntoView(true);
-    this.getMerchant();
+
+    this.quantity = 1;
+    this.getDish();
+
   }
 
-  getMerchant() {
+  getCurrentUserCart() {
+    this.cartService.getCurrentUserCart().subscribe(
+      (response) => this.cart = response as Cart
+    );
+  }
+
+  getDish() {
     this.dishService.getById(this.id).subscribe(
       (response) => {
-        const dish = response as Dish;
-        this.merchant = dish.merchant;
+        this.dish = response as Dish;
+        this.dishImage = 'http://localhost:8080/image/' + this.dish.image;
+        this.calculateDishTotal();
       }
     );
   }
 
-  refreshNavbar($event) {
-    this.refreshNum = $event;
+  increaseQuantity() {
+    this.quantity++;
+    this.calculateDishTotal();
+  }
+
+  decreaseQuantity() {
+    if (this.quantity > 1) {
+      this.quantity--;
+      this.calculateDishTotal();
+    }
+  }
+
+  calculateDishTotal() {
+    this.dishTotal = this.dish.price * this.quantity;
+  }
+
+  addToCart() {
+    if (!this.authService.isLoggedIn()) {
+      document.getElementById('app-navbar-customer').scrollIntoView(true);
+      this.notificationService.showErrorMessage('Hãy đăng nhập để có thể đặt món');
+      return;
+    }
+    const cartDetail = {
+      dish: this.dish,
+      quantity: this.quantity
+    };
+    this.cartService.addDishToCart(cartDetail).subscribe(
+      {
+        next: (response) => {
+          this.cart = response as Cart;
+          this.notificationService.showSuccessMessage(`Đã thêm vào giỏ hàng </br> (${this.dish.name}) x ${this.quantity}`);
+        },
+        error: (error) => {
+          this.notificationService.showErrorMessage('Có lỗi xảy ra!');
+          console.log(error.error.message);
+        }
+      }
+    );
   }
 }
