+++ src/app/merchant/food-list/food-list.component.ts
@@ -1,8 +1,4 @@
 import { Component, OnInit } from '@angular/core';
-import {Dish} from '../../model/dish';
-import {DishService} from '../../service/dish/dish.service';
-import {User} from '../../model/user';
-import {AuthService} from '../../service/auth/auth.service';
 
 @Component({
   selector: 'app-food-list',
@@ -10,25 +6,10 @@ import {AuthService} from '../../service/auth/auth.service';
   styleUrls: ['./food-list.component.css']
 })
 export class FoodListComponent implements OnInit {
-  dishes: Dish[] = [];
-  user: User = {};
-  userId: number;
 
-  constructor(private dishService: DishService,
-              private authService: AuthService) {
-    this.authService.currentUser.subscribe(value => this.user = value);
-  }
+  constructor() { }
 
   ngOnInit() {
-    this.getAllMerchantDishes();
-  }
-
-  getAllMerchantDishes() {
-    this.userId = this.authService.getCurrentUserId();
-    this.dishService.getAllMerchantDishes(this.userId).subscribe(dishesFromBE => {
-      this.dishes = dishesFromBE;
-      console.log(this.dishes);
-    });
   }
 
 }
