+++ src/app/merchant/merchant-routing.module.ts
@@ -12,10 +12,6 @@ const routes: Routes = [
   {
     path: 'edit',
     component: MerchantEditComponent
-  },
-  {
-    path: ':id/edit',
-    component: FoodListComponent
   }
 ];
