+++ src/app/merchant/merchant.module.ts
@@ -7,6 +7,7 @@ import {MerchantEditComponent} from './merchant-edit/merchant-edit.component';
 import {ReactiveFormsModule} from '@angular/forms';
 import {FoodListComponent} from './food-list/food-list.component';
 
+
 @NgModule({
   declarations: [MerchantEditComponent, FoodListComponent],
   imports: [
