+++ src/app/service/dish/dish.service.ts
@@ -37,8 +37,4 @@ export class DishService {
   getMostPurchasedDishes() {
     return this.httpClient.get(`${API_URL}/dishes/most-purchased/8`);
   }
-
-  getAllMerchantDishes(userId: number): Observable<Dish[]> {
-    return this.httpClient.get<Dish[]>(`${API_URL}/merchants/user/${userId}/merchant/dishes`);
-  }
 }
