+++ src/app/shared/navbar-customer/navbar-customer.component.ts
@@ -1,24 +1,21 @@
-import {Component, Input, OnChanges, OnInit, SimpleChanges} from '@angular/core';
+import {Component, OnInit} from '@angular/core';
 import {AuthService} from '../../service/auth/auth.service';
 import {CartDetail} from '../../model/cart-detail';
 import {CartService} from '../../service/cart/cart.service';
 import {Cart} from '../../model/cart';
+import {User} from '../../model/user';
 
 @Component({
   selector: 'app-navbar-customer',
   templateUrl: './navbar-customer.component.html',
   styleUrls: ['./navbar-customer.component.css']
 })
-export class NavbarCustomerComponent implements OnInit, OnChanges {
-
-  @Input() refreshNum;
+export class NavbarCustomerComponent implements OnInit {
 
   currentUser: any;
   loggedIn: boolean;
   cart: CartDetail[] = [];
   total: number;
-
-  isCartEmpty: boolean;
 
   constructor(private authService: AuthService,
               private cartService: CartService
@@ -46,12 +43,7 @@ export class NavbarCustomerComponent implements OnInit, OnChanges {
     this.cartService.getCurrentUserCart().subscribe(
       (response) => {
         this.cart = (response as Cart).cartDetails;
-        this.isCartEmpty = this.cart.length === 0;
       }
     );
-  }
-
-  ngOnChanges(changes: SimpleChanges): void {
-      this.getCart();
   }
 }
